dafi-web3/src/tests/unit/ContractService.test.ts
import { ContractService } from '@/services/ContractService';
import { Principal } from '@dfinity/principal';

describe('ContractService', () => {
  let contractService: ContractService;
  const mockIdentity = {};
  const mockPrincipal = Principal.fromText('2vxsx-fae');

  beforeEach(() => {
    contractService = new ContractService(mockIdentity);
  });

  describe('Asset Management', () => {
    it('should tokenize an asset', async () => {
      const assetData = {
        owner: mockPrincipal,
        metadata: {
          name: 'Test Asset',
          description: 'Test Description',
          location: 'Test Location',
          type: 'Test Type',
        },
        totalShares: BigInt(1000),
        pricePerShare: BigInt(100),
      };

      const mockResponse = {
        status: 'success',
        assetId: '123',
      };

      // @ts-ignore - Mock implementation
      contractService.assetActor = {
        tokenizeAsset: jest.fn().mockResolvedValue(mockResponse),
      };

      const result = await contractService.tokenizeAsset(assetData);
      expect(result).toEqual(mockResponse);
    });

    it('should get asset details', async () => {
      const mockAsset = {
        owner: mockPrincipal,
        metadata: {
          name: 'Test Asset',
          description: 'Test Description',
          location: 'Test Location',
          type: 'Test Type',
        },
        totalShares: BigInt(1000),
        pricePerShare: BigInt(100),
      };

      // @ts-ignore - Mock implementation
      contractService.assetActor = {
        getAssetDetails: jest.fn().mockResolvedValue(mockAsset),
      };

      const result = await contractService.getAssetDetails('123');
      expect(result).toEqual(mockAsset);
    });
  });

  describe('Investment Management', () => {
    it('should create an investment', async () => {
      const mockResponse = {
        status: 'success',
        investmentId: '456',
      };

      // @ts-ignore - Mock implementation
      contractService.investmentActor = {
        invest: jest.fn().mockResolvedValue(mockResponse),
      };

      const result = await contractService.invest('123', BigInt(500));
      expect(result).toEqual(mockResponse);
    });

    it('should get investments by asset', async () => {
      const mockInvestments = [{
        investor: mockPrincipal,
        assetId: '123',
        amount: BigInt(500),
        timestamp: BigInt(Date.now()),
      }];

      // @ts-ignore - Mock implementation
      contractService.investmentActor = {
        getInvestmentsByAsset: jest.fn().mockResolvedValue(mockInvestments),
      };

      const result = await contractService.getInvestmentsByAsset('123');
      expect(result).toEqual(mockInvestments);
    });
  });

  describe('Return Management', () => {
    it('should create a return', async () => {
      const returnData = {
        assetId: '123',
        amount: BigInt(1000),
        distributionDate: BigInt(Date.now()),
      };

      const mockResponse = {
        status: 'success',
        returnId: '789',
      };

      // @ts-ignore - Mock implementation
      contractService.returnActor = {
        createReturn: jest.fn().mockResolvedValue(mockResponse),
      };

      const result = await contractService.createReturn(returnData);
      expect(result).toEqual(mockResponse);
      expect(result.status).toBe('success');
      expect(result.returnId).toBeDefined();
    });

    it('should get returns by asset', async () => {
      const mockReturns = [{
        assetId: '123',
        amount: BigInt(1000),
        distributionDate: BigInt(Date.now()),
      }];

      // @ts-ignore - Mock implementation
      contractService.returnActor = {
        getReturnsByAsset: jest.fn().mockResolvedValue(mockReturns),
      };

      const result = await contractService.getReturnsByAsset('123');
      expect(result).toEqual(mockReturns);
    });
  });
});
